tiny refactor
